+++ src/docs/components/code-block.svelte
@@ -35,7 +35,7 @@
 	{/if}
 </button>
 
-<style>
+<style lang="postcss">
 	[data-rehype-pretty-code-fragment] :global(pre) {
 		font-weight: initial !important;
 		@apply border-neutral-700/50 bg-neutral-800/50;