+++ src/lib/builders/combobox/types.ts
@@ -1,4 +1,4 @@
-import type { BuilderReturn } from '$lib/internal/types.js';
+import type { BuilderReturn, WhenTrue } from '$lib/internal/types.js';
 import type { ChangeFn } from '$lib/internal/helpers/index.js';
 import type { Writable } from 'svelte/store';
 import type { createCombobox } from './create.js';
@@ -10,7 +10,17 @@ export type ComboboxOption<Value> = {
 	label?: string;
 };
 
-export type CreateComboboxProps<Value> = {
+export type ComboboxSelected<Multiple extends boolean, Value> = WhenTrue<
+	Multiple,
+	ComboboxOption<Value>[],
+	ComboboxOption<Value>
+>;
+
+export type CreateComboboxProps<
+	Value,
+	Multiple extends boolean = false,
+	S extends ComboboxSelected<Multiple, Value> = ComboboxSelected<Multiple, Value>
+> = {
 	/**
 	 * Options for positioning the popover menu.
 	 *
@@ -61,19 +71,19 @@ export type CreateComboboxProps<Value> = {
 	 *
 	 * @default undefined
 	 */
-	defaultSelected?: ComboboxOption<Value>;
+	defaultSelected?: S;
 
 	/**
 	 * An optional controlled store that manages the selected option of the combobox.
 	 */
-	selected?: Writable<ComboboxOption<Value> | undefined>;
+	selected?: Writable<S>;
 
 	/**
 	 * A change handler for the selected store called when the selected would normally change.
 	 *
 	 * @see https://melt-ui.com/docs/controlled#change-functions
 	 */
-	onSelectedChange?: ChangeFn<ComboboxOption<Value> | undefined>;
+	onSelectedChange?: ChangeFn<S | undefined>;
 
 	/**
 	 * Whether or not to close the combobox menu when the user clicks
@@ -115,6 +125,8 @@ export type CreateComboboxProps<Value> = {
 	 * @default false
 	 */
 	forceVisible?: boolean;
+
+	multiple?: Multiple;
 };
 
 export type ComboboxItemProps<Value> = ComboboxOption<Value> & {
@@ -124,8 +136,32 @@ export type ComboboxItemProps<Value> = ComboboxOption<Value> & {
 	disabled?: boolean;
 };
 
-export type Combobox = BuilderReturn<typeof createCombobox>;
-export type ComboboxElements = Combobox['elements'];
-export type ComboboxOptions = Combobox['options'];
-export type ComboboxStates = Combobox['states'];
-export type ComboboxHelpers = Combobox['helpers'];
+export type Combobox<
+	Value = unknown,
+	Multiple extends boolean = false,
+	S extends ComboboxSelected<Multiple, Value> = ComboboxSelected<Multiple, Value>
+> = BuilderReturn<typeof createCombobox<Value, Multiple, S>>;
+
+export type ComboboxElements<
+	Value = unknown,
+	Multiple extends boolean = false,
+	S extends ComboboxSelected<Multiple, Value> = ComboboxSelected<Multiple, Value>
+> = Combobox<Value, Multiple, S>['elements'];
+
+export type ComboboxOptions<
+	Value = unknown,
+	Multiple extends boolean = false,
+	S extends ComboboxSelected<Multiple, Value> = ComboboxSelected<Multiple, Value>
+> = Combobox<Value, Multiple, S>['options'];
+
+export type ComboboxStates<
+	Value = unknown,
+	Multiple extends boolean = false,
+	S extends ComboboxSelected<Multiple, Value> = ComboboxSelected<Multiple, Value>
+> = Combobox<Value, Multiple, S>['states'];
+
+export type ComboboxHelpers<
+	Value = unknown,
+	Multiple extends boolean = false,
+	S extends ComboboxSelected<Multiple, Value> = ComboboxSelected<Multiple, Value>
+> = Combobox<Value, Multiple, S>['helpers'];