+++ src/lib/builders/tooltip/create.ts
@@ -3,7 +3,6 @@ import {
 	addMeltEventListener,
 	builder,
 	createElHelpers,
-	derivedVisible,
 	effect,
 	executeCallbacks,
 	generateId,
@@ -22,8 +21,7 @@ import {
 
 import { useFloating, usePortal } from '$lib/internal/actions/index.js';
 import type { MeltActionReturn } from '$lib/internal/types.js';
-import { onMount, tick } from 'svelte';
-import { get, writable } from 'svelte/store';
+import { derived, get, writable, type Writable } from 'svelte/store';
 import type { TooltipEvents } from './events.js';
 import type { CreateTooltipProps } from './types.js';
 
@@ -40,10 +38,14 @@ const defaults = {
 	portal: 'body',
 	closeOnEscape: true,
 	disableHoverableContent: false,
+	group: undefined,
 } satisfies CreateTooltipProps;
 
 type TooltipParts = 'trigger' | 'content' | 'arrow';
 const { name } = createElHelpers<TooltipParts>('tooltip');
+
+// Store a global map to get the currently open tooltip.
+const groupMap = new Map<string | true, Writable<boolean>>();
 
 export function createTooltip(props?: CreateTooltipProps) {
 	const withDefaults = { ...defaults, ...props } satisfies CreateTooltipProps;
@@ -59,29 +61,31 @@ export function createTooltip(props?: CreateTooltipProps) {
 		portal,
 		closeOnEscape,
 		disableHoverableContent,
+		group,
 	} = options;
 
 	const openWritable = withDefaults.open ?? writable(withDefaults.defaultOpen);
 	const open = overridable(openWritable, withDefaults?.onOpenChange);
 
-	const activeTrigger = writable<HTMLElement | null>(null);
+	type OpenReason = 'pointer' | 'focus';
+	const openReason = writable<null | OpenReason>(null);
 
 	const ids = {
 		content: generateId(),
 		trigger: generateId(),
-	};
+	} as const;
 
 	let clickedTrigger = false;
 
-	onMount(() => {
-		if (!isBrowser) return;
-		activeTrigger.set(document.querySelector(`[aria-describedby="${ids.content}"]`));
-	});
+	const getEl = (part: keyof typeof ids) => {
+		if (!isBrowser) return null;
+		return document.getElementById(ids[part]);
+	};
 
 	let openTimeout: number | null = null;
 	let closeTimeout: number | null = null;
 
-	function openTooltip() {
+	function openTooltip(reason: OpenReason) {
 		if (closeTimeout) {
 			window.clearTimeout(closeTimeout);
 			closeTimeout = null;
@@ -90,6 +94,8 @@ export function createTooltip(props?: CreateTooltipProps) {
 		if (!openTimeout) {
 			openTimeout = window.setTimeout(() => {
 				open.set(true);
+				// Don't override the reason if it's already set.
+				openReason.update((prev) => prev ?? reason);
 				openTimeout = null;
 			}, get(openDelay));
 		}
@@ -101,11 +107,19 @@ export function createTooltip(props?: CreateTooltipProps) {
 			openTimeout = null;
 		}
 
-		if (isBlur && isMouseInTooltipArea) return;
+		if (isBlur && isMouseInTooltipArea) {
+			// Normally when blurring the trigger, we want to close the tooltip.
+			// The exception is when the mouse is still in the tooltip area.
+			// In that case, we have to set the openReason to pointer, so that
+			// it can close when the mouse leaves the tooltip area.
+			openReason.set('pointer');
+			return;
+		}
 
 		if (!closeTimeout) {
 			closeTimeout = window.setTimeout(() => {
 				open.set(false);
+				openReason.set(null);
 				if (isBlur) clickedTrigger = false;
 				closeTimeout = null;
 			}, get(closeDelay));
@@ -116,6 +130,7 @@ export function createTooltip(props?: CreateTooltipProps) {
 		returned: () => {
 			return {
 				'aria-describedby': ids.content,
+				id: ids.trigger,
 			};
 		},
 		action: (node: HTMLElement): MeltActionReturn<TooltipEvents['trigger']> => {
@@ -132,7 +147,7 @@ export function createTooltip(props?: CreateTooltipProps) {
 				}),
 				addMeltEventListener(node, 'pointerenter', (e) => {
 					if (isTouch(e)) return;
-					openTooltip();
+					openTooltip('pointer');
 				}),
 				addMeltEventListener(node, 'pointerleave', (e) => {
 					if (isTouch(e)) return;
@@ -143,7 +158,7 @@ export function createTooltip(props?: CreateTooltipProps) {
 				}),
 				addMeltEventListener(node, 'focus', () => {
 					if (clickedTrigger) return;
-					openTooltip();
+					openTooltip('focus');
 				}),
 				addMeltEventListener(node, 'blur', () => closeTooltip(true)),
 				addMeltEventListener(node, 'keydown', (e) => {
@@ -164,7 +179,9 @@ export function createTooltip(props?: CreateTooltipProps) {
 		},
 	});
 
-	const isVisible = derivedVisible({ open, activeTrigger, forceVisible });
+	const isVisible = derived([open, forceVisible], ([$open, $forceVisible]) => {
+		return $open || $forceVisible;
+	});
 
 	const content = builder(name('content'), {
 		stores: [isVisible, portal],
@@ -185,34 +202,34 @@ export function createTooltip(props?: CreateTooltipProps) {
 			let unsubPortal = noop;
 
 			const unsubDerived = effect(
-				[isVisible, activeTrigger, positioning, portal],
-				([$isVisible, $activeTrigger, $positioning, $portal]) => {
-					if (!$isVisible || !$activeTrigger) {
+				[isVisible, positioning, portal],
+				([$isVisible, $positioning, $portal]) => {
+					const triggerEl = getEl('trigger');
+					if (!$isVisible || !triggerEl) {
 						unsubPortal();
 						unsubFloating();
 						return;
 					}
-					tick().then(() => {
-						const floatingReturn = useFloating($activeTrigger, node, $positioning);
-						unsubFloating = floatingReturn.destroy;
-						if (!$portal) {
-							unsubPortal();
-							return;
+
+					const floatingReturn = useFloating(triggerEl, node, $positioning);
+					unsubFloating = floatingReturn.destroy;
+					if (!$portal) {
+						unsubPortal();
+						return;
+					}
+					const portalDest = getPortalDestination(node, $portal);
+					if (portalDest) {
+						const portalReturn = usePortal(node, portalDest);
+						if (portalReturn && portalReturn.destroy) {
+							unsubPortal = portalReturn.destroy;
 						}
-						const portalDest = getPortalDestination(node, $portal);
-						if (portalDest) {
-							const portalReturn = usePortal(node, portalDest);
-							if (portalReturn && portalReturn.destroy) {
-								unsubPortal = portalReturn.destroy;
-							}
-						}
-					});
+					}
 				}
 			);
 
 			const unsubEvents = executeCallbacks(
-				addMeltEventListener(node, 'pointerenter', openTooltip),
-				addMeltEventListener(node, 'pointerdown', openTooltip)
+				addMeltEventListener(node, 'pointerenter', () => openTooltip('pointer')),
+				addMeltEventListener(node, 'pointerdown', () => openTooltip('pointer'))
 			);
 
 			return {
@@ -240,16 +257,30 @@ export function createTooltip(props?: CreateTooltipProps) {
 
 	let isMouseInTooltipArea = false;
 
-	effect([isVisible, activeTrigger], ([$isVisible, $activeTrigger]) => {
-		if (!$isVisible || !$activeTrigger) return;
+	effect(open, ($open) => {
+		if (!$open) return;
+
+		const groupValue = get(group);
+		if (groupValue === undefined || groupValue === false) {
+			return;
+		}
+
+		// Close the currently open tooltip in the same group
+		// and set this tooltip as the open one.
+		const currentOpen = groupMap.get(groupValue);
+		currentOpen?.set(false);
+		groupMap.set(groupValue, open);
+	});
+
+	effect([open, openReason], ([$open, $openReason]) => {
+		if (!$open || !isBrowser) return;
 		return executeCallbacks(
 			addEventListener(document, 'mousemove', (e) => {
-				const contentEl = document.getElementById(ids.content);
-				if (!contentEl) return;
+				const contentEl = getEl('content');
+				const triggerEl = getEl('trigger');
+				if (!contentEl || !triggerEl) return;
 
-				const polygonElements = get(disableHoverableContent)
-					? [$activeTrigger]
-					: [$activeTrigger, contentEl];
+				const polygonElements = get(disableHoverableContent) ? [triggerEl] : [triggerEl, contentEl];
 				const polygon = makeHullFromElements(polygonElements);
 
 				isMouseInTooltipArea = pointInPolygon(
@@ -260,11 +291,10 @@ export function createTooltip(props?: CreateTooltipProps) {
 					polygon
 				);
 
-				if (
-					isMouseInTooltipArea ||
-					(document.activeElement === $activeTrigger && !clickedTrigger)
-				) {
-					openTooltip();
+				if ($openReason !== 'pointer') return;
+
+				if (isMouseInTooltipArea) {
+					openTooltip('pointer');
 				} else {
 					closeTooltip();
 				}